+++ src/graphics.h
@@ -68,7 +68,7 @@ void oshu_draw_path(struct oshu_display *display, struct oshu_path *path);
  * Actually draws two more-or-less parallel lines. Follow the curve nicely but
  * might make ugly loops when the path is too... loopy.
  */
-void oshu_draw_thick_path(struct oshu_display *display, struct oshu_path *path, int width);
+void oshu_draw_thick_path(struct oshu_display *display, struct oshu_path *path, double width);
 
 /**
  * Draw a regular polyline that should look like a circle.
